feat(IdentityService): add identity service and use it for catalog en…

+++ src/Services/Catalog/Catalog.API/Controllers/CategoriesController.cs
@@ -1,7 +1,11 @@
 using Catalog.API.Models;
+using Catalog.API.Services.Interfaces;
 using Catalog.Core.Entities;
 using Catalog.Core.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Primitives;
+using Shared.Models;
 
 namespace Catalog.API.Controllers
 {
@@ -10,10 +14,12 @@ namespace Catalog.API.Controllers
 	public class CategoriesController : ControllerBase
 	{
 		private readonly ICategoryService categoryService;
+		private readonly IAuthService authService;
 
-		public CategoriesController(ICategoryService service)
+		public CategoriesController(ICategoryService service, IAuthService authService)
 		{
 			categoryService = service;
+			this.authService = authService;
 		}
 
 		/// <summary>
@@ -21,6 +27,7 @@ namespace Catalog.API.Controllers
 		/// </summary>
 		/// <returns>A list of categories</returns>
 		[HttpGet()]
+		[AllowAnonymous]
 		public async Task<IEnumerable<Category>> GetCategories()
 		{
 			var categories = await categoryService.ListCategoriesAsync();
@@ -37,6 +44,9 @@ namespace Catalog.API.Controllers
 		[HttpPost("{id}")]
 		public async Task<IActionResult> AddCategory(int id, UpdatedCategory category)
 		{
+			IActionResult? result = await this.AuthorizeUser(UserPermissions.Create);
+			if (result != null) return result;
+
 			var addedCategory = await categoryService.AddCategoryAsync(id, category.Name, category.ImageUrl, category.ParentCategoryId);
 
 			return Created($"api/categories/{id}", addedCategory);
@@ -51,6 +61,9 @@ namespace Catalog.API.Controllers
 		[HttpPut("{id}")]
 		public async Task<IActionResult> UpdateCategory(int id, UpdatedCategory category)
 		{
+			IActionResult? result = await this.AuthorizeUser(UserPermissions.Update);
+			if (result != null) return result;
+
 			var updatedCategory = await categoryService.UpdateCategoryAsync(id, category.Name, category.ImageUrl, category.ParentCategoryId);
 
 			return Ok(updatedCategory);
@@ -64,10 +77,30 @@ namespace Catalog.API.Controllers
 		[HttpDelete("{id}")]
 		public async Task<IActionResult> DeleteCategory(int id)
 		{
+			IActionResult? result = await this.AuthorizeUser(UserPermissions.Delete);
+			if (result != null) return result;
+
 			var category = await categoryService.GetCategoryAsync(id);
 			await categoryService.DeleteCategoryAsync(category);
 
 			return NoContent();
+		}
+
+		private async Task<IActionResult?> AuthorizeUser(UserPermissions permission)
+		{
+			this.HttpContext.Request.Headers.TryGetValue("Authorization", out StringValues headerValue);
+			if (string.IsNullOrEmpty(headerValue)) return Unauthorized();
+
+			bool isAuthorized = await this.authService.ValidateUserToken(headerValue, permission);
+			if (!isAuthorized) return Problem(
+				type: "/docs/errors/forbidden",
+				title: "Authenticated user is not authorized.",
+				detail: "Token is not valid or user does not have rights to access the endpoint.",
+				statusCode: StatusCodes.Status403Forbidden,
+				instance: HttpContext.Request.Path
+			);
+
+			return null;
 		}
 	}
 }
